backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
pydantic==2.4.2
python-dotenv==1.0.0
firebase-admin==6.2.0
efficientnet-pytorch==0.7.1
scikit-learn==1.3.1
pandas==2.1.1
timm==0.9.2